Connector for databricks volume storage

The ConnectorDatabricksVolume class, built on top of connector::connector class. It is a file storage connector for accessing and manipulating files inside Databricks volumes.

Super classes

connector::Connector -> connector::ConnectorFS -> ConnectorDatabricksVolume

Active bindings

path

character Path to the file storage on Volume

catalog

character Databricks catalog

schema

character Databricks schema

full_path

character Full path to the file storage on Volume

Methods

Public methods

ConnectorDatabricksVolume$new()

ConnectorDatabricksVolume$clone()

Inherited methods

connector::Connector$list_content_cnt()

connector::Connector$print()

connector::Connector$read_cnt()

connector::Connector$remove_cnt()

connector::Connector$write_cnt()

connector::ConnectorFS$create_directory_cnt()

connector::ConnectorFS$download_cnt()

connector::ConnectorFS$download_directory_cnt()

connector::ConnectorFS$remove_directory_cnt()

connector::ConnectorFS$tbl_cnt()

connector::ConnectorFS$upload_cnt()

connector::ConnectorFS$upload_directory_cnt()

Method new()

Initializes the connector for Databricks volume storage.

Usage

ConnectorDatabricksVolume$new(
  full_path = NULL,
  catalog = NULL,
  schema = NULL,
  path = NULL,
  extra_class = NULL,
  force = FALSE,
  ...
)

Arguments

full_path

character Full path to the file storage in format catalog/schema/path. If NULL, catalog, schema, and path must be provided.

catalog

character Databricks catalog

schema

character Databricks schema

path

character Path to the file storage

extra_class

character Extra class to assign to the new connector.

force

logical If TRUE, the volume will be created without asking if it does not exist.

...

Additional arguments passed to the initialize method of superclass

Returns

A new ConnectorDatabricksVolume object

Method clone()

The objects of this class are cloneable with this method.

Usage

ConnectorDatabricksVolume$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Create Volume file storage connector
cnt <- ConnectorDatabricksVolume$new(full_path = "catalog/schema/path")

cnt

# List content
cnt$list_content_cnt()

# Write to the connector
cnt$write_cnt(iris, "iris.rds")

# Check it is there
cnt$list_content_cnt()

# Read the result back
cnt$read_cnt("iris.rds") |>
  head()
}