Write content to the connector

Addition write methods for databricks connectors implemented for connector::write_cnt():

ConnectorDatabricksTable: Creates temporary volume to write object as a parquet file and then convert it to a table.

ConnectorDatabricksVolume: Reuses the connector::write_cnt() method for ConnectorDatabricksVolume, but always sets the catalog and schema as defined in when initializing the connector.

Usage

write_cnt(
  connector_object,
  x,
  name,
  overwrite = zephyr::get_option("overwrite", "connector"),
  ...
)

# S3 method for class 'ConnectorDatabricksTable'
write_cnt(
  connector_object,
  x,
  name,
  overwrite = FALSE,
  ...,
  method = "volume",
  tags = NULL
)

# S3 method for class 'ConnectorDatabricksVolume'
write_cnt(connector_object, x, name, overwrite = FALSE, ...)

Arguments

connector_object

Connector The connector object to use.

x

The object to write to the connection

name

character Name of the content to read, write, or remove. Typically the table name.

overwrite

Overwrite a file if it already exists

...

ConnectorDatabricksVolume: Additional parameters to pass to the brickster::db_volume_write() method

method

ConnectorDatabricksTable: Which method to use for writing the table. Options:

volume - using temporary volume to write data and then convert it to a table.

tags

Named list containing tag names and tag values, e.g. list("tag_name1" = "tag_value1", "tag_name2" = "tag_value2") More info here

Value

ConnectorDatabricksVolume object